Make playbooks self-executable
Add shebang to each playbook to be able to run them like

> ./run-something.yml

instead of

> ansible-playbook run-something.yml

Reason to do that is that these playbooks are used so frequently that
even avoiding few symbols at run them will save some noticeable amount
of time. Previously I had an alias

alias apl='ansible-playbook'

for this but this solution is even shorter, so I decided to switch to
it.

Some documentation was added also to each playbook.